+++ GeoCAT-comp-examples/comp_only/moc_globe_atl_example.py
@@ -47,7 +47,7 @@ tlat = ds.TLAT[:].values.astype(np.double)
 
 # Read important parameters from input data
 nyaux = lat_aux_grid.shape[0]  # 395
-km = np.max(kmt.values).astype(np.int)
+km = np.max(kmt.values).astype(np.int32)
 ny = tarea.shape[0]
 nx = tarea.shape[1]
